feat(farm): separate farm groups
